\name{GHQp-package}
\alias{GHQp-package}
\alias{GHQp}
\docType{package}
\title{Gaussian Hermite Quadrature with prunning}
\description{This package has a function that is used to obtain quadrature points to approximate an integral applying the pruning approach to eliminate some points that contribute little to the approximation.}
\author{
Freddy Hernandez (maintainer) <fhernanb@gmail.com>
}
\keyword{
package
}

